+++ src/main/java/br/com/devsuperior/eventos/entities/Categoria.java
@@ -14,7 +14,7 @@ public class Categoria implements Serializable {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+  private Integer id;
 
   private String descricao;
 
@@ -23,16 +23,16 @@ public class Categoria implements Serializable {
 
   public Categoria() {}
 
-  public Categoria(Long id, String descricao) {
+  public Categoria(Integer id, String descricao) {
     this.id = id;
     this.descricao = descricao;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
